+++ app/src/main/java/com/tianyu/seelove/ui/fragment/ManageFragment.java
@@ -3,6 +3,8 @@ package com.tianyu.seelove.ui.fragment;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
+import android.os.Handler;
+import android.os.Message;
 import android.support.v4.app.Fragment;
 import android.util.DisplayMetrics;
 import android.view.LayoutInflater;
@@ -10,17 +12,27 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.GridView;
+import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
+import android.widget.Toast;
 
+import com.nostra13.universalimageloader.core.ImageLoader;
 import com.tianyu.seelove.R;
 import com.tianyu.seelove.adapter.VideoGridAdapter;
+import com.tianyu.seelove.common.MessageSignConstant;
+import com.tianyu.seelove.controller.UserController;
+import com.tianyu.seelove.model.entity.user.SLUser;
 import com.tianyu.seelove.model.entity.video.VideoInfo;
+import com.tianyu.seelove.ui.activity.base.BaseActivity;
 import com.tianyu.seelove.ui.activity.user.MyInfoActivity;
+import com.tianyu.seelove.ui.activity.user.UserInfoActivity;
 import com.tianyu.seelove.ui.activity.video.VideoListActivity;
+import com.tianyu.seelove.utils.ImageUtil;
 import com.tianyu.seelove.utils.LogUtil;
 import com.tianyu.seelove.view.dialog.CustomProgressDialog;
+import com.tianyu.seelove.view.dialog.PromptDialog;
 import com.tianyu.seelove.wxapi.QQEntryActivity;
 import com.tianyu.seelove.wxapi.WXEntryActivity;
 
@@ -32,9 +44,17 @@ import java.util.ArrayList;
  * @author shisheng.zhao
  * @date 2017-03-29 15:03
  */
-public class ManageFragment extends Fragment implements View.OnClickListener {
+public class ManageFragment extends Fragment implements View.OnClickListener, Handler.Callback {
     ArrayList<VideoInfo> videoInfos;
     public CustomProgressDialog customProgressDialog;
+    public PromptDialog promptDialog;
+    private View view = null;
+    private UserController controller;
+    public Handler handler;
+    private String userName = "天宇";
+    private Long userId;
+    private ImageView headView;
+    private TextView titleView, nameView;
 
     @Override
     public void onAttach(Activity activity) {
@@ -46,13 +66,28 @@ public class ManageFragment extends Fragment implements View.OnClickListener {
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         LogUtil.d("ManageFragment____onCreate");
+        handler = new Handler(this);
+        promptDialog = new PromptDialog(getActivity());
+        controller = new UserController(getActivity(), handler);
     }
 
     @Override
     public View onCreateView(final LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         LogUtil.d("ManageFragment____onCreateView");
-        View view = inflater.inflate(R.layout.fragment_manage, container, false);
-        TextView titleView = (TextView) view.findViewById(R.id.titleView);
+        // 防止onCreateView被多次调用
+        if (null != view) {
+            ViewGroup parent = (ViewGroup) view.getParent();
+            if (null != parent)
+                parent.removeView(view);
+        } else {
+            view = inflater.inflate(R.layout.fragment_manage, container, false);
+            initView(view);
+        }
+        return view;
+    }
+
+    private void initView(View view) {
+        titleView = (TextView) view.findViewById(R.id.titleView);
         RelativeLayout userInfoLayout = (RelativeLayout) view.findViewById(R.id.userInfoLayout);
         LinearLayout videoLayout = (LinearLayout) view.findViewById(R.id.videoLayout);
         userInfoLayout.setOnClickListener(new View.OnClickListener() {
@@ -94,14 +129,16 @@ public class ManageFragment extends Fragment implements View.OnClickListener {
         gridView.setStretchMode(GridView.NO_STRETCH);
         gridView.setNumColumns(size); // 设置列数量=列表集合数
         gridView.setAdapter(new VideoGridAdapter(getActivity(), videoInfos));
-        return view;
-    }
-
-    private void initView(View view) {
         Button qqLoginBtn = (Button) view.findViewById(R.id.qqLoginBtn);
         Button wechatLoginBtn = (Button) view.findViewById(R.id.wechatLoginBtn);
+        Button registBtn = (Button) view.findViewById(R.id.registBtn);
+        Button loginBtn = (Button) view.findViewById(R.id.loginBtn);
+        headView = (ImageView) view.findViewById(R.id.headView);
+        nameView = (TextView) view.findViewById(R.id.nameView);
         qqLoginBtn.setOnClickListener(this);
         wechatLoginBtn.setOnClickListener(this);
+        registBtn.setOnClickListener(this);
+        loginBtn.setOnClickListener(this);
     }
 
     @Override
@@ -122,9 +159,76 @@ public class ManageFragment extends Fragment implements View.OnClickListener {
                 startActivityForResult(intent, 0);
                 break;
             }
+            case R.id.registBtn: {
+                customProgressDialog = new CustomProgressDialog(getActivity(), getString(R.string.loading));
+                customProgressDialog.show();
+                controller.create(userName, "我是从微信返回的字段");
+                break;
+            }
+            case R.id.loginBtn: {
+                customProgressDialog = new CustomProgressDialog(getActivity(), getString(R.string.loading));
+                customProgressDialog.show();
+                controller.login(userId, "", "");
+                break;
+            }
             default:
                 break;
         }
+    }
+
+    /**
+     * Handler发送message的逻辑处理方法
+     *
+     * @param msg
+     * @return
+     */
+    @Override
+    public boolean handleMessage(Message msg) {
+        if (customProgressDialog != null)
+            customProgressDialog.dismiss();
+        if (promptDialog == null || promptDialog.isShowing())
+            promptDialog = new PromptDialog(getActivity());
+        SLUser user;
+        String code;
+        String message;
+        switch (msg.what) {
+            case MessageSignConstant.USER_CREATE_SUCCESS:
+                user = (SLUser) msg.getData().getSerializable("user");
+                userId = user.getUserId();
+                titleView.setText(user.getNickName());
+                nameView.setText("创建token:" + user.getToken4RongCloud());
+                Toast.makeText(getActivity(), "创建成功：" + user.toString(), Toast.LENGTH_LONG).show();
+                break;
+            case MessageSignConstant.USER_CREATE_FAILURE:
+                code = msg.getData().getString("code");
+                message = msg.getData().getString("message");
+                promptDialog.initData(getString(R.string.user_create_failure), message);
+                promptDialog.show();
+                break;
+            case MessageSignConstant.USER_LOGIN_SUCCESS:
+                user = (SLUser) msg.getData().getSerializable("user");
+                titleView.setText(user.getNickName());
+                nameView.setText("登录获取token:" + user.getToken4RongCloud());
+                ImageLoader.getInstance().displayImage(user.getHeadUrl(), headView, ImageUtil.getImageOptions());
+                Toast.makeText(getActivity(), "登录成功：" + user.toString(), Toast.LENGTH_LONG).show();
+                break;
+            case MessageSignConstant.USER_LOGIN_FAILURE:
+                code = msg.getData().getString("code");
+                message = msg.getData().getString("message");
+                promptDialog.initData(getString(R.string.user_login_failure), message);
+                promptDialog.show();
+                break;
+            case MessageSignConstant.SERVER_OR_NETWORK_ERROR:
+                promptDialog.initData("", msg.getData().getString("message"));
+                promptDialog.show();
+                break;
+            case MessageSignConstant.UNKNOWN_ERROR:
+                promptDialog.initData("", getString(R.string.unknown_error));
+                promptDialog.show();
+                break;
+        }
+        // 加载效果关闭
+        return false;
     }
 
     @Override
