+++ app/src/main/res/layout/activity_userinfo.xml
@@ -26,35 +26,6 @@
                 android:layout_gravity="center_horizontal"
                 android:orientation="horizontal">
 
-                <ImageView
-                    android:id="@+id/headView"
-                    android:layout_width="50dp"
-                    android:layout_height="50dp" />
-
-                <Button
-                    android:id="@+id/testBtn1"
-                    android:layout_width="wrap_content"
-                    android:layout_height="40dp"
-                    android:background="@color/white"
-                    android:text="1模拟注册"
-                    android:textColor="@color/red_7a" />
-
-                <Button
-                    android:id="@+id/testBtn2"
-                    android:layout_width="wrap_content"
-                    android:layout_height="40dp"
-                    android:layout_marginLeft="20dp"
-                    android:background="@color/white"
-                    android:text="2模拟自登陆更新"
-                    android:textColor="@color/red_7a" />
-            </LinearLayout>
-
-            <LinearLayout
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center_horizontal"
-                android:orientation="horizontal">
-
                 <TextView
                     android:id="@+id/name"
                     android:layout_width="wrap_content"
